fix: Đổi lại database id là Bigint

+++ MedicalAppointmentBookingSystem/src/main/java/org/mabs/service/MedicalRecordService.java
@@ -5,5 +5,5 @@ import org.mabs.dto.MedicalRecordResponse;
 import java.util.List;
 
 public interface MedicalRecordService {
-    List<MedicalRecordResponse> getMedicalRecordsByPatient(Integer patientId);
+    List<MedicalRecordResponse> getMedicalRecordsByPatient(Long patientId);
 }
